Speed up program by only using numpy
I found that I was able to get a massive speed increase (15-20 times) by using only numpy. This optimization was necessary for me as I was using this implementation every bar of data during backtests. Instead of reorganizing my code to be more like yours I decided to just paste my code in.